Add resources configuration options to all jobs
This patch adds resources configuration options to all job pods.

Some production grade clusters force to set resources to be able to calculate
quotas correctly. So it is crucial to be able to set those resources an all
pods.

This patch include breaking changes, because it changes the structure of the
values.

Closes #2

Signed-off-by: Sven Haardiek <sven.haardiek@uni-muenster.de>